+++ azure-storage-queue/ChangeLog.md
@@ -5,3 +5,4 @@
 ## Version 1.0.0:
 
 - The package has switched from Apache 2.0 to the MIT license.
+- Queue messages can now have an arbitrarily large or infinite time to live